krishimitra/krishimitra/requirements.txt
Django==5.2.6
python-decouple==3.8
Pillow==10.0.1
whitenoise==6.6.0
requests==2.32.3


